Move search bar into data plugin (#35389)
* Moved filter bar to data plugin (still not working)

* Moved filter options from search bar into filter bar
