+++ src/legacy/core_plugins/data/public/index_patterns/index_patterns_service.ts
@@ -51,7 +51,6 @@ import setupRouteWithDefaultPattern from 'ui/index_patterns/route_setup/load_def
 import { getFromSavedObject, isFilterable } from 'ui/index_patterns/static_utils';
 
 // IndexPattern, StaticIndexPattern, StaticIndexPatternField, Field
-
 import * as types from 'ui/index_patterns';
 
 /**
