+++ src/legacy/core_plugins/data/public/search_bar/components/index.tsx
