+++ src/legacy/core_plugins/data/public/search_bar/components/search_bar.tsx
@@ -28,8 +28,7 @@ import { FilterBar } from 'ui/filter_bar';
 import { IndexPattern } from 'ui/index_patterns';
 import { Storage } from 'ui/storage';
 
-import { data } from 'plugins/data';
-const { QueryBar } = data.query.ui;
+import { QueryBar } from '../../query_bar';
 
 interface Query {
   query: string;
@@ -124,16 +123,16 @@ class SearchBarUI extends Component<Props, State> {
 
   public render() {
     const filtersAppliedText = this.props.intl.formatMessage({
-      id: 'common.ui.searchBar.filtersButtonFiltersAppliedTitle',
+      id: 'data.search.searchBar.filtersButtonFiltersAppliedTitle',
       defaultMessage: 'filters applied.',
     });
     const clickToShowOrHideText = this.state.isFiltersVisible
       ? this.props.intl.formatMessage({
-          id: 'common.ui.searchBar.filtersButtonClickToShowTitle',
+          id: 'data.search.searchBar.filtersButtonClickToShowTitle',
           defaultMessage: 'Select to hide',
         })
       : this.props.intl.formatMessage({
-          id: 'common.ui.searchBar.filtersButtonClickToHideTitle',
+          id: 'data.search.searchBar.filtersButtonClickToHideTitle',
           defaultMessage: 'Select to show',
         });
