+++ src/legacy/core_plugins/data/public/search_bar/index.tsx
@@ -17,6 +17,4 @@
  * under the License.
  */
 
-import './directive';
-
-export { SearchBar } from './components';
+export { SearchBarService } from './search_bar_service';
