+++ src/legacy/ui/public/filter_bar/filter_bar.tsx
@@ -33,9 +33,9 @@ import classNames from 'classnames';
 import React, { Component } from 'react';
 import chrome from 'ui/chrome';
 import { IndexPattern } from 'ui/index_patterns';
-import { FilterOptions } from 'ui/search_bar/components/filter_options';
 import { FilterEditor } from './filter_editor';
 import { FilterItem } from './filter_item';
+import { FilterOptions } from './filter_options';
 
 const config = chrome.getUiSettingsClient();
