+++ tsconfig.json
@@ -12,7 +12,7 @@
       ],
       "test_utils/*": [
         "src/test_utils/public/*"
-      ]
+      ],
     },
     // Support .tsx files and transform JSX into calls to React.createElement
     "jsx": "react",
