+++ x-pack/plugins/translations/translations/zh-CN.json
